Add files via upload
This is e-commerce sales data analysis with thousands of entries in multiple tables. This project analyses the data and answers different questions related to it. 
The data was collected from Kaggle and imported into the MySQL Workbench. The tech stack used were Python, MySQL and Python libraries (NumPy, Pandas, Matplotlib, Seaborn). The tools used were Jupyter Notebook and MySQL Workbench.